requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-session==0.7.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2 